agent-sdk/typescript: move host-side source via git mv
Phase 2 of the TS SDK split. Relocates the host-side source (and the
test files that cover it) to the new sibling package. Imports inside
the moved files still use the old caller-relative paths — Phase 3
rewrites them to `@synadia-ai/agents`. The tree therefore does NOT
typecheck after this commit; that is expected and corrects in
Phase 3.

Moves (`git mv`, blame preserved):

  client-sdk/typescript/src/service.ts                    →  agent-sdk/typescript/...
  client-sdk/typescript/src/stream/chunk-encoder.ts       →  agent-sdk/typescript/...
  client-sdk/typescript/src/testing/reference-agent.ts    →  agent-sdk/typescript/...
  client-sdk/typescript/src/testing/index.ts              →  agent-sdk/typescript/...
  client-sdk/typescript/test/unit/chunk-encoder.test.ts            →  agent-sdk/typescript/...
  client-sdk/typescript/test/unit/heartbeat-builder.test.ts        →  agent-sdk/typescript/...
  client-sdk/typescript/test/integration/agent-service.test.ts     →  agent-sdk/typescript/...

`heartbeat/payload.ts` is split inline:
- caller keeps `HeartbeatPayload` + `decodeHeartbeatPayload` + the
  `KNOWN_FIELDS` set (decoder side stays caller-local).
- a NEW `agent-sdk/typescript/src/heartbeat/payload.ts` holds
  `BuildHeartbeatPayloadOptions`, `buildHeartbeatPayload`, and
  `encodeHeartbeatPayload`. It imports `HeartbeatPayload` and
  `AgentSubject` from `@synadia-ai/agents` so the type lives in one
  place.

`heartbeat-payload.test.ts` only exercises the decoder, so it stays
caller-side untouched. The temporary `_placeholder.test.ts` from
Phase 1 is deleted now that real test files exist on the host side.
