+++ agent-sdk/typescript/src/stream/chunk-encoder.ts
